Use shared LTO properties for CFI
Rather than explicitly adding LTO flags for CFI, the CFI santizer can use the
LTO property to enable LTO for modules that use CFI. This allows multiple
features to use LTO and not add duplicate -flto and --plugin options to the
build (which llvm-ar rejects). Ensure we don't try to enable CFI and LTO for
dependencies with LTO explicitly disabled.

Test: make -j32 PRODUCT-aosp_x86-eng checkbuild tests

Change-Id: I61f83642d5ec58e69f8dcef3854c10b3d751f0f9
Signed-off-by: mydongistiny <jaysonedson@gmail.com>